# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CreatorOS API
  version: 0.10.0
  description: REST contract for the CreatorOS modular monolith.
servers:
  - url: http://localhost:8080
    description: Local API
security: []
...
paths:
...
  /api/v1/orders/{order_id}/thread:
    get:
      operationId: getOrderConversationThread
      summary: Get or initialize the conversation thread for an order
      tags: [Communication]
      security:
        - sessionCookie: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Conversation thread with participants and recent messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationThreadResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/threads/{thread_id}/messages:
    get:
      operationId: listThreadMessages
      summary: List messages in a conversation thread
      tags: [Communication]
      security:
        - sessionCookie: []
      parameters:
        - name: thread_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: since
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: List of messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: sendThreadMessage
      summary: Post a message into a conversation thread
      tags: [Communication]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - name: thread_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/notifications:
    get:
      operationId: listNotifications
      summary: List in-app notifications for the authenticated user
      tags: [Communication]
      security:
        - sessionCookie: []
      parameters:
        - name: unread_only
          in: query
          required: false
          schema: { type: boolean, default: false }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Paginated notification list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/notifications/unread-count:
    get:
      operationId: getUnreadNotificationCount
      summary: Get count of unread in-app notifications
      tags: [Communication]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Unread count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadNotificationCountResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/notifications/{notification_id}/read:
    post:
      operationId: markNotificationRead
      summary: Mark a single in-app notification as read
      tags: [Communication]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - name: notification_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Notification marked as read.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      summary: Mark all in-app notifications as read
      tags: [Communication]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204":
          description: All notifications marked as read.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/user/notification-preferences:
    get:
      operationId: getNotificationPreferences
      summary: Get communication and notification preferences
      tags: [Communication]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Notification preferences.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    put:
      operationId: updateNotificationPreferences
      summary: Update communication and notification preferences
      tags: [Communication]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferencesRequest"
      responses:
        "200":
          description: Updated preferences.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/communication/notifications/stream:
    get:
      operationId: streamNotifications
      summary: Server-Sent Events stream for real-time unread counts and alerts
      tags: [Communication]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: SSE event stream.
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthenticated"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: creatoros_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Must match the creatoros_csrf cookie bound to the current session.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The JSON body is malformed, too large, contains unknown fields, or is not a single object.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthenticated:
      description: A valid active session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ForbiddenAction:
      description: The authenticated role or permission cannot perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ConversationParticipant:
      type: object
      required: [thread_id, user_id, last_read_at, joined_at]
      properties:
        thread_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_display_name:
          type: string
          example: Jane Doe
        last_read_at:
          type: string
          format: date-time
        joined_at:
          type: string
          format: date-time
    Message:
      type: object
      required: [id, thread_id, sender_user_id, body, created_at]
      properties:
        id:
          type: string
          format: uuid
        thread_id:
          type: string
          format: uuid
        sender_user_id:
          type: string
          format: uuid
        sender_name:
          type: string
          example: Alex Brand
        body:
          type: string
          example: Hello creator, looking forward to the shoot!
        created_at:
          type: string
          format: date-time
    ConversationThread:
      type: object
      required: [id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ConversationParticipant"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    SendMessageRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 5000
          example: Hello, please check the lighting requirements for tomorrow.
    Notification:
      type: object
      required: [id, user_id, kind, title, body, action_url, is_read, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            - order_update
            - submission_received
            - revision_requested
            - submission_approved
            - payment_received
            - escrow_released
            - payout_update
            - campaign_invitation
            - campaign_response
            - new_message
            - system
        title:
          type: string
          example: New message received
        body:
          type: string
          example: Hello, please check the lighting requirements for tomorrow.
        action_url:
          type: string
          example: /orders/00000000-0000-0000-0000-000000000001
        is_read:
          type: boolean
          example: false
        read_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    NotificationPreferences:
      type: object
      required: [user_id, email_notifications, order_updates, messages, updated_at]
      properties:
        user_id:
          type: string
          format: uuid
        email_notifications:
          type: boolean
          example: true
        order_updates:
          type: boolean
          example: true
        messages:
          type: boolean
          example: true
        updated_at:
          type: string
          format: date-time
    UpdateNotificationPreferencesRequest:
      type: object
      properties:
        email_notifications:
          type: boolean
        order_updates:
          type: boolean
        messages:
          type: boolean
    ConversationThreadResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ConversationThread"
    MessageResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Message"
    MessageListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    NotificationListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
          example: 12
    UnreadNotificationCountResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [count]
          properties:
            count:
              type: integer
              example: 3
    NotificationPreferencesResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/NotificationPreferences"
...